server/requirements.txt
Flask==2.0.2
mysql-connector-python==8.0.26
Flask-Cors==3.0.10